Fix nested closeout scanner edge cases
